Fix: equalsAndHashCode 기준 필드 변경
