+++ src/main/java/com/flab/dduikka/common/domain/Auditable.java
@@ -1,7 +1,6 @@
 package com.flab.dduikka.common.domain;
 
 import java.time.LocalDateTime;
-import java.util.Objects;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
@@ -14,20 +13,5 @@ public abstract class Auditable {
 
 	protected Auditable(LocalDateTime createdAt) {
 		this.createdAt = createdAt;
-	}
-
-	@Override
-	public boolean equals(Object o) {
-		if (this == o)
-			return true;
-		if (o == null || getClass() != o.getClass())
-			return false;
-		Auditable auditable = (Auditable)o;
-		return Objects.equals(createdAt, auditable.createdAt);
-	}
-
-	@Override
-	public int hashCode() {
-		return Objects.hash(createdAt);
 	}
 }
