+++ src/main/java/com/flab/dduikka/domain/vote/domain/VoteRecord.java
@@ -1,6 +1,7 @@
 package com.flab.dduikka.domain.vote.domain;
 
 import java.time.LocalDateTime;
+import java.util.Objects;
 
 import com.flab.dduikka.common.domain.Auditable;
 
@@ -20,10 +21,6 @@ public class VoteRecord extends Auditable {
 
 	private Boolean isCanceled;
 
-	private VoteRecord(LocalDateTime createdAt) {
-		super(createdAt);
-	}
-
 	@Builder
 	public VoteRecord(Long voteRecordId, Long voteId, Long userId, VoteType voteType, boolean isCanceled,
 		LocalDateTime createdAt) {
@@ -33,6 +30,23 @@ public class VoteRecord extends Auditable {
 		this.userId = userId;
 		this.voteType = voteType;
 		this.isCanceled = isCanceled;
+	}
+
+	@Override
+	public boolean equals(Object o) {
+		if (this == o)
+			return true;
+		if (o == null || getClass() != o.getClass())
+			return false;
+		VoteRecord that = (VoteRecord)o;
+		return Objects.equals(voteRecordId, that.voteRecordId) && Objects.equals(voteId, that.voteId)
+			&& Objects.equals(userId, that.userId) && voteType == that.voteType && Objects.equals(
+			isCanceled, that.isCanceled);
+	}
+
+	@Override
+	public int hashCode() {
+		return Objects.hash(voteRecordId, voteId, userId, voteType, isCanceled);
 	}
 
 	public void cancel() {
